Clarify addresses of IMSIC interrupt files, and support major subdivision of address space.
Add recommendations for how to arrange the memory regions of IMSIC
interrupt files in a machine's address space.  Also, modify an
Advanced PLIC's MSI address configuration registers (mmsiaddrcfg,
mmsiaddrcfgh, smsiaddrcfg, smsiaddrcfgh) to support the subdivision
of the machine's address space so that different groups of harts are
constrained to separate parts of the address space (usually because
the harts are on separate chips in a "multi-socket" configuration).